# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/api",
    "paths": {
        "/images": {
            "get": {
                "description": "images list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "HandleImagesList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Image"
                            }
                        }
                    }
                }
            }
        },
        "/images/download": {
            "get": {
                "description": "download images from links file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "HandleDownloadImages",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/images/upload": {
            "post": {
                "description": "upload images",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "HandleImagesUpload",
                "parameters": [
                    {
                        "type": "file",
                        "description": "images",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/images/{id}": {
            "get": {
                "description": "downoald an image from list",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "HandleDownloadImage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Image": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "download_date": {
                    "type": "string"
                },
                "file_extension": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "local_name": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}